Folder Reorganization/ Video Functionality Updated
Updated the Video Playback Functionality so there is only 2 pairs of media players and media textures. I also put in a boolean for Looped? in the data table. I tried setting it up so that if you wanted to only have the video play once instead of looping you could just change the boolean appropriately. However, only Looped seems to be working. With the boolean set to false, for some reason the video plays fine when you go up to it (currently Manhole is set to false), but when you step away and come back, none of the UI shows back up. Not even the text or anything, but when you leave the collider the phase out animation shows up. Not sure whats going on there. TBD