Updated Input Mappings/ Particle Effects Updated
All hardcoded keys have been made into Input Actions and added to the Default Input Mappings we have assigned to the GameController. I also messed around in the JetFlame Niagara Particle we had and set up collision, changed up the retribution on the collision settings and randomized the collion normals. You can see the difference when you press L in gameplay by the car. A block was added to test collision. AI's were also placed next to the firetruck the same as the design and I cleaned up more of the blueprint code and folder structure.