Boundary & Transitional UI Updated
Low Level Functionality for Navigational Bounds is setup up. When the player gets close to the bounds area. A material will dissolve in showing a wall blocking their path. When the player moves away from it, it dissolves away. This material has an open color variable that can be changed in editor and can dissolve as quickly or slowly as we desire it to. Right now it's set up to dissolve within 1 second each way. Transitional UI Design has been updated along with minor low level functionality. More work to add to that tomorrow.