Jetflame Updated
All 4 phases are functional and integrated into the Residential Houses Level. Press L to navigate to each different phase. After the 4th phase, it turns off everything and sets the index back to zero so it can be repeated ad infinitum if desired. Easy to be retranslated to different logic should we decide to change up the way it is called. Scales, Locations, and SpawnRates are open variables that can be changed in editor.