Interaction System Updated
There is a bug where the window only does the change mesh function during the fence interaction, instead of the window interaction. Not exactly sure why this is happening. Everything else seems to be working cohesively. This system probably needs a once over with some paired programming to get to it's final workable stage but at the most it's only an hour or two of work left. After this is the substep setup part of the interaction which we will need to go over again as well.