Xray Vision Functionality Added
Majority of XRay Vision Functionality is complete. All colors and FX are vector/scalar parameters that can be adjusted dynamically or through the editor instance itself (XRayMat_Inst). Material added to XRay PostProcessVolume added to ResidentialHouses Level. Set Array Value to 1 to test out the Post Process Material. To test out any static mesh, click on the static mesh and check the Custom Render Depth Pass to True and change the Custom Static value to the appropriate one (1,2,3) and it should glow via the appropriate color. These are already tested out on some static meshes in the level (couple of trees and the firetruck). I added 3 different highlight colors to point out objects of different meaning (green, red, and blue). There is also a blending filter I'm finalizing that should make it look much better than it already does and give it a nice FX look when playing in the level sequence we are using this for. I should be able to finish it off tomorrow and then get on to the other stuff.