Xray Execution Updated
XRay Vision can now be called on flip flop through the M key. I can make it go faster or slower, and all of the color variables are interchangeable. Change the custom stencil values from 1-3 on static meshes in the level to set them to different colors. I'll go ahead and make a level sequence for zooming through the window to the products on the shelving in the garage and it will automatically update if we change meshes or positioning of things. Xray vision will also update because of the way I will be calling it in the Interactive Actor. Give it a try when you get a chance. And be sure to test out the actual X Ray part (move behind walls and obstructions to continue seeing the highlighted objects). Let me know what you think.