Interact Updated
Finished the Xray Highlight Level Sequence Setup. For some reason only some of the static meshes that have values >= 1 and have render custom depth pass as true highlight. Not all of them. And I'm not sure exactly why. Feel free to test it out yourself if you get a chance. Decided to move past that and finished the other interaction you wanted. Created a believable lever, created dynamic functionality inside of it so if we needed to manipulate it outside of the level sequence, we could. I also made the level sequence and hooked up all of the things that we needed to make it work. Feel free to give it a try when you get a chance. When I start next week I will begin work on the spline stuff with the firehose and stuff. Have a good weekend man.