Spline Created/Setup
Firehose Spline has been created. Open variables are prioritized around Section Length (what is the size of the static mesh on the x axis), color, ability to scale or not, full mesh or no, and I have the connections set up for the attaching to the proper parts on the firetruck and segment through there. Next is finalization for simple visual calls (hooked to fire hydrant, laying on the ground, and attached to actor) and then work on the level sequences.