Action/FP Call to FHose Interactive Constructed
Once the Truck is moved a new interactable actor is spawned on top of the firehose with the necessary variables for the player to trigger should they enter its path. UI pops up once the FP enters its collision however Pressing F doesn't seem to call, so I will have to work on some booleans (maybe pair programming would help as I may be overcomplicating the functionality). Action button is functionally filled out, however the visibility option still has to be choreographed. Need to think of the best way to call it (probaly make a branch seeing if FHose is available). Skeleton functionality of this is already choreographed I just need to fine tune it and test it out. All best to have this functionality looked over together to make sure it is cohesive and sound. Should be finished easily once a good systems route has been decided.