ResidentialHouses Updates
Civilian Models have been updated to 5 different models. These can be changed via an open variable Enum on the Blueprint. This will automatically change the skeletal mesh and the animation blueprint associated with it. All models are set up with Idle only.

More Asset Fillers have been added to the Garage and the Sedan's transform has been centered. The garage's back wall has also been altered to include a backdoor leading to the backyard.

Collision Boxes have been added to the House's front and back entrance, Pawns can no longer enter the residence. I also added modeled wires with the same material as the powerlines and connected the rest of the powerlines together, as well as the houses within the player's navigation. Everything should be hooked up and looking uniform now.