Interactive Updates & Cutscenes
I got the On-Complete Functionality setup so that the player can interact with the fire truck interactive (now placed on the fire hydrant across the street from the house) as soon as they enter the game. Once the player watches the cutscene and all of the hose meshes are visible the player can then go to the hose handle and interact with it to play the fade sequence with an extinguishing sound. Further functionality would simply turn the fire off on the jetflame, assuming that the Jetflame is turned on to the first phase at the beginning of the level?? If you do want it that way then I will need some sort of an idea on "cue events" of when the other phases are supposed to fire off. I sort of figured we could have it on the smoke phase until the garage is opened and then after a certain amount of time the other phases would fire off unless the hose is used. I was thinking dividing them up until the last phase at 5 minutes? But I can make that an open variable if you want so you can change it to whatever. Let me know what you think of all that.

Also I added a simple interactive that plays a level sequence of the door opening and closing, except I didn't use a camera so the action just happens. Hope that's cool. It was simple and effective. But yeah you can get in the backyard now through the garage through gameplay.