Updated Thermal/Gas Meter UI Functionality
Design is complete for Thermal and Gas Meter UI (can be changed if you don't like it). ThermalVision functionality is now limited to the garage only. Keypress was made into an Input Action and added to the Default. SceneTextures/Mats and Cube Mesh created and functional during gameplay. Post Processing Mat (thermalvision) still needs to be replicated on the texture. Calling and Collapsing still needs to be set up for the Gas Meter.