ThermalGun Functionality Updated
Integrated LineTrace for Objects with the ThermalGun. Now whenever the Thermalgun is being used it will detect object with the Object Type Heat Objects and will update the new UI text on the Thermalgun. Test it out on the car and the t.v. in the garage.  (Fyi, for some reason the tv isn't showing up on the thermalgun even though I checked the custom render depth pass and raised the stencil value) but it's merely meant to show the functionality of the thermalgun so it shouldn't effect us currently.