Level Updates and Unit Action UI Functionality Updated
Changed up the architecture of the side fence area and created the necessary level sequences for both fence gates to be called on by FPSPawn and AI. All cutscenes are updated and made for AI as well. Datatables have been updated for Action and DescriptiveActors(.ie Checklist). Functionality should call for the necessary Action Buttons to be called and checklists to be updated contingent upon whether those checklist items have been completed or not. All Action Button Collapse Functionality is set for beginning use but needs residual true statement completed for once the action has been completed so that the action button becomes collapsed and no longer available for reuse.