Controls UI 90% Complete
Controls UI Design & Functionality is pretty much completely finished. UI Appears when the gamehud starts up. I have the controls for Movement (WASD), Camera|Select (Mouse Input), Drone Mode (C), Unit Call Commands (Tab), Thermal Camera *Garage Only* (M), Gas Meter (G), Checklist (K), and Walkie (F). However, I need to include F for Interact and I don't like the way the Mouse Input Image I made came out and need to change the color scheme to make it a little more uniform so I will update that tomorrow and include Zoom, since it's sort of an easy to miss feature.

UI consists of a Title Text with a variety of controls text with images of the buttons associated with them. The colors/opacity is binded to functions I made to see if those keys are being pressed. There are three different colors that are tied up to the UI text variables (disabled, yellow, and white on the colors struct). If the player is not in the garage then the Title text of the Thermal Gun and the M text on the image are set to the disabled color. Once in the garage they will reset to white (default) and light up to yellow on the colors struct the same as the rest.