Controls UI Updated and Descripters Updated UI Work has begun
More trial work on highlight. Will more than likely need two pairs of eyes in order to finalize it. Updated the icon images and have everything else set up and good to go for the controls ui. Started work on updating the descriptor UI. Changed up the design and currently setting up functionality so it can be tested out. Should more than likely go over it together to see what should be done next. A few finalizing touches need to be made before it is finished on the design perspective.