Descriptive UI functionality Updated
Setup Collision Functionality and Back Button Functionality. For some reason it is stuck showing the Residential Manhole variable on the data table on all of the Descriptive Actors. ID Tags are different and all code originating from the Training Tracker to the moment it is supposed to be called to the screen has been backchecked.