Finalized Gas Meter UI Functionality
Gas Meter is now functional. When the player presses G (Shift is now used for the Helmet) the Gas Meter comes out and Input is focused on UI only with the mouse cursor. Default values will show for the O2, CO2, and the checkmark for Comb|Explos (unchecked). A linetrace is also placed right after the call to scan the environment for Gas Locations (object collision, same as Heat Objects). If the player didn't scan a Gas Location object then it will read specific values for the buttons pressed (O2, CO2, and Comb). The backdoor of the garage has been changed to a Gas Location Object.