Gamehud/Splash/Thermal&GasGunVisibility Updated
Added Version Text to Main Menu and GameHUD in the lower left corner of the screen. Added some images made up for the Splash Screen and Editor Screen and assigned them. Put in code to call for visibility of thermal and gas meter guns whenever the player is using them. This way other characters in the world space will be able to visually see the guns in their hands while they are being used. Animation is not included.