Finalized Dispatch UI
UI is visible at start up (after gamehud has replaced the mainmenu) and when the residential houses level loads it begins the Text Dialogue Animations. Everything phases in and out fairly well and the new plugin works perfectly. I have it set to 20 characters and 0.5s for the punctuality delay (whenever a comma or period appears in the text said pause occurs). The dialogue text is called through an interface with the BPI GameController so right now I just have it set up with the latin lingo.