Dispatch UI updated. Meshes Updated. Textures Updated. Visibility Updated
Dispatch UI now only appears in the Scenario Level Stream, Static Meshes of the Radio, Gas Meter, and Thermal Gun have been added to the FPS Pawn toolbelt and work with it's animations as a parent/child socket relationship. Visibility is updated so that when the user is utilizing the tool it will flash to their hands and then back to their toolbelt when they are finished with it. This will only be visible to other players but it's a nice touch.