Shooting Gallery Updated
Updated the Niagara Fluid being shown for Spraying. I also updated the parent/child hierarchy to be attached to the FPS Camera. Everything is rotated and functional the way that it should be. I also updated the names of the meshes incorporated for the Spraygun and moved them to the correct Mesh Folder in the Art Section.