EVFireSprayLevel Added/Updated
Added a new Level Stream (EVFireSpray) to the Residential Houses Level. In the Level Blueprint it is set up so that if the player selects a module with the word "Spray" in it, then it will load the EVFireSpray Level Stream. For development purposes, I have the booleans set up so it automatically loads up the EVFireSpray if you just load up from the ResidentialHouses Level instead of going through the MainMenu. Functionality should be good to go. The EVFire actor now has a trigger box that checks to see if the component that triggers it has a tag for FireKiller. If it does, then it increments an integer variable (TimesHit) and plays the spray audio. When the spray exits the trigger box it resets the do once node and the player can hit the fire again and increment and continue the functionality. Once the process has been repeated to 4 times, the next time it is hit it plays the same audio, but all of the niagra fluid particles are deactivated, the TimesHit variable is reset back to 0 and the actor itself is destroyed.