FireHoseOptions UI
Created the FireHoseOptions UI. Set up the functionality so it fades in when called and presents two options to the player. Clicking on each button will fade out the UI, set the UI to collapsed in the GameHud, sets all input modes back to game only and hides the mouse cursor. After that it will play the appropriate level sequence that will lay down the hose in whatever position they chose. After that a boolean is marked true or false on whether it is in the correct position or not (found in the GameController) and then gives mobility back to the player.