More Updates
Changed Up functionality so it goes straight to hydrant then the hose on the truck and you decide what option you want. If you make the wrong one. It lays out but you can go back up to the hose on the truck and change it to the front one. Level Sequences still need to be assigned. updated the Adjust Position Level Sequence to involve correct actors hidden variables. So it just needs to be called. Added another trigger actor for spraying when the player is ready for it. Placing it in the right location and minimal overlap functionality is setup.