Ending UI Updated/ FireHose Options Updated
Added additional functionality so it can be used at the nozzle as well. Changed up the design to include actions/locations/and a double check option. All buttons should be functional. I also updated the Ending UI and have most of it's booleans set up for proper visual display of the available text. Nothing is tied to tick and it should have a couple more checks before it is ready for testing. 90% finished. Also updated a few functions in the gamecontroller.