Scenario 1 Finalization
Created an Interactive Wench Actor with similar functionality as the Interactive Hose. Collision is turned on after the fire is put out. When the uses the F key near the Wench Interactive (back of the truck) a fade sequence plays and the car is pulled out to the driveway. Functionality is set up with the fire to call it and then dismiss it, but Yusuf should finish that part as he set up the fire and I don't want to mess with his tag system he set up. I'll lead him through what needs to be done to finalize the second fire after it is pulled out. But 10 seconds after the fade sequence the Ending UI shows up on the screen and the final results are tallied. I also created a simple Game Time Recorder Function that isn't dependant on Tick. So the Playtime should be accurate. As long as they don't play over an hour lol. Overall Distinctions Text is set up with sample text. And the Electric Panel Text will always read as incorrect since we didn't have enough time to finalize that functionality in the level (another couple of hours of work). Other than that. The player can get from the start of the level to Scenario completion. We just need to finalize the bug with the Training Tracker and it should be ready for a build.