Resolved Text Bugs and UI
Fixed Color Coding on EndingScreen for ALL conditions of pass and fail. Should be correct now. I also noticed some left over buttons on the left hand corner of the screen when the gas meter is being used so I collapsed them so they wouldn't be on the viewport anymore.