Comment Updates
Gas Meter is now Toggleable and player pawn has correct movement parameters during UI interaction. Thermal Gun/Gas Gun cannot be used while the other is being used. Neither Gun UIs can be seen while the player is in drone mode (even in the garage). At the beginning of the game Lowest Settings are automatically initiated and Lowest Setting button is disabled to show that it is chosen. All Settings can still be changed and updated via the Options Menu. I also took out the F in the controls UI but added functionality calls for Scroll to only show up in the Controls UI when the player is in drone mode. During normal gameplay the Scroll Control UI is collapsed.