Familiarization Ending Conditions Updated
UI Created, And Functionality set up for Checklist to be updated everytime you interact with a descriptive actor. A counter is also incremented and once it reaches it's maximum there is functionality to call the Ending Screen for the Familiarization Level. The only thing that needs to be resolved is an issue with logic pertaining to getting the checklist to check it's boxes. Once that is resolved it should be functional.