Ending UI Updated
Added Functionality for the Talked to Owner Boolean. There is now a section that checks to see if the player has talked to the owner. If they have then it will it will have the appropriate text below it and be the appropriate color of text. All available Ending conditions have been updated. Level Sequence for the Garage Door Explosion has been prepped and is ready for completion.