Garage Door Explosion Functionality
Level Sequence Created, Functionality is assigned with a random bool with weight. If the player goes up to the door and hasn't taken care of the fire yet the player will rather immediately enter the endscreen UI or (based on weight) will turn the Niagara particle auto activate to true and then play the level sequence of the garage door explosion. After the level sequence plays the Endscreen UI appears.