Last Change Updates
Adjusted the Left Gate in all Level Sequences to Match up what is in the level. The shot image wasn't trailed out in the Speak to Victim sequence so I fixed it, it should play the whole level sequence now (8 seconds). I changed out the white car for the tow truck and updated the Towing Level Sequence. I also updated the ESS Fire Management so it should be displayable along with Yusuf's fire. I also added functionality in the Interactive Actors when it plays a level sequence to set the Owner No See appropriately. I also noticed a bit of functionality that made the pawn invisible during that functionality so I separated it. I also took another look at the functionality for the ending conditions. And the reason there was nothing happening after you pressed the M key after the second time you put out the fire is because there was nothing attached to the true statement lol. So I added the call to the EndScreen UI and updated the values to the text on it.