Tuesday Night Updates after Beta Review
All Text References to Thermal Gun have been changed to "Thermal Imaging Camera". Thermal Imaging Camera is now visualized with the T Keypress instead of M. Additional Animations were made to the Chris Civilian/Victim so he looks more agitated and moves around a little in Idle. I also updated the cutscene for the VictimTalk to show at the new position at the front of the house, however the firefighter's animation won't move for some reason (currently debugging). I also disconnected everything pertaining to the Gas Meter and got rid of the text in the EndScreen UI and the Controls UI/Data Tables. Objects in the Garage have been resized for realism. An additional Interactive for Tools was created and is placed at the Front Fire Engine right beside where the Hose Options is located. The Hydrant Interactive's Collision starts out with No Collision but after the Tools Interactive has been completed then it will be active and normal gameplay will proceed. A simple fade and pickup sound plays for the Interactive Tool Interaction and only Activates once. I also made the Thermal Imaging Camera itself a parent of the FirstPersonCamera on the FPSPawn so, you should now be able to pivot the Thermal Imaging Camera up and down instead of just left and right. If anything is missing from the viewport, the objects themselves will need LineTrace Object Titles attributed to them tos how up as special on the viewport.