Wednesday Night Updates
Added 2 cascade fires (ESS and EV) and 2 Startup Smoke Cascade particles associated with the EV. I've added a Get and Set for a Bool named "Realism?" It just needs to be set up within options. The Level Functionality should default to the cascade fire  events replacing the niagara fluid ones. The cascade smoke and ESS Fire will start off in the beginning just like the original one. 90 seconds later it will check to see if the fire has been put out, if it hasn't the second fire on the car will go off as well as the turbo jet fire (which is visible on their computers). After the fire has been put out on the car all fires and smokes associated with it go out and the same goes for the ESS spray functionality. It's ready for playtesting. I also added a simple function check for the drone mode whenever you go to drone mode, it should check if the thermal imaging camera is being used, if it is it will phase out,  if it isn't it won't do anything.