Added Arrows to Interactives
Blue Scaled Arrows were added to each interactive in the Scenario Level and are positioned to point downwards to the Interactive UI that will pop up on the screen once the player gets close enough. When the player enters the collision zone, the arrow will disappear, and will reappear when the exit the collision zone. If the player interacts with the interactable then the arrow will not return. This last part of functionality is different with the Spray hoses as they can be interacted with multiple times in different locations so that is a simple appear/disappear functionality tied to the collision zone only.