Spray Steps Updated, Gas Explosion Level Sequence Updated
Player should have to spray the door now before they can access the backdoor of the garage. Functionality should be complete for 10 sprays and then the collision for the interactive for the backdoor changes to query only instead of no collision and you should be able to complete the regular events of the Scenario to completion. Gas Explosion Level sequence was updated with a new pawn with new animation and debugged gas explosion niagara particle appearance and set it to play at 100% of the time when the player tries to enter the front garage door before the EVFire has been put out.