Updated Logic on Fence
Updated FireHoseOptions to not show LeftGate Hint if it isn't opened. When BackYard option is chosen it checks to see if it has been opened. If it isn't, it opens it. If it already has been it goes to normal functionality. I also updated the GateOpener logic to check to see if the LeftGate has been opened when the player contacts the Trigger. If it has then it ignores the player. If it hasn't then it goes to normal functionality. That way nothing goes off if the player opens the gate through the firehoseoptions.