Last Friday Updates
New Firehose Mesh was created for the garage. It is set to hidden currently and has no collision. A simple level sequence of .10 seconds turns on the visibility and it is set to play after you close the Thermal Imaging Camera after successfully utilizing it after you take out the EV Fire a second time in the driveway. I also altered the WenchPull Level Sequence to not alter the relative locaiton of the ESS Trigger once the car gets moved. This way the car can be shot at once it is moved and the ESS can be sprayed again at it's static location in the garage. The Level Loader was updated to check a function called House Update. I basically just checked to see if realism was true or not and if it is, then the Houses and Background Houses Stream Level are opened. If it isn't true then nothing will happen and the streaming levels will remain at their default mode (which should be off). I also changed the Thermal Imaging Checkbox to a text value and added a new Struct, Enum, and Data Table for the Ending Screen UI. It is not finished yet, but the majority of its guts have been strewed together.