Burst Update 1
Created a Cracked/Fully Open Door Level Sequence. Created Door Action UI. Door Action UI has two button options (Fully Open Door and Crack Open Door). The same visual aesthetics that align with the FireHose Options are the same with this UI. Back Button works similarly. Choosing either option will disable the button from further use. Each button will fade out the Door Options UI and collapse it and then play the appropriate level sequence attributed to it (fully, cracked). After the level sequence plays the player is able to move around again. *More functionality will be added to disable the trigger for the Door Actions once the player has Fully Opened the Door. The EndScreen UI has also been updated for the new Data Table structure.

The Thermal Gun UI now has a dark blurry backdrop and the opening animation is slightly different but takes the same amount of time to appear on the screen. There is also a Thermal Notes Text that can appear like Hints that will pop up on the top of the screen whenever we need to call it.

I also updated the DT_TrainingInteractive and took out the BGarageDoor and the Gates and added a new one for Electrical Box that will be placed next to the front door. It is set to a normal fade and sound sequence and will only activate once and is mostly used to change the boolean from false to true for the EndingUI.