Backdoor Edits
Finalized Trigger Functionality and Spray Functionality for Cracked/Full Open Back Door. Ready for Playtesting. Spray won't work if the backdoor hasn't been interacted with. If the player uses the full open option it will play through normally with fully opening the door, turning on the EVFire and ESS Spray Triggers for Spray Collision and spraying the car instead of the door and won't update the Cracked Door Boolean. If the player selects Cracked Door it will crack the door slightly and turn on the backdoor spray trigger and the player can then spray the in the door frame and after it is sprayed to completion, the door will automatically open the rest of the way and the backdoor open boolean will be marked as true as well as turning on the triggers for the EVFire and ESS for oncoming Spray Collision. Further functionality with spray will continue on to spraying the EVFire and so on and so forth.