ESS Second Spray Functionality Update
Majority of functionality is set up and is ready for playtesting. Will more than likely need a couple of tweaks but it is oriented around the specifications that were asked for. I also updated the audio on the gas explosion and began it earlier than it was originally. Thermal Gun UI is also updated with the new text.