Commander Vehicle Parent/Child Relationship Updated
Following John's tutorial I changed the Parent/Child relationship of the commander vehicle to the BP_TrainingStep_Interactive and replaced original components to the parent's. I updated the idtag, filled out extra spots for the commander vehicle and the backdoor to the DT_TrainingStepInteractive and set the Run Parent Overlap to false. Ready for playtesting.