Parent/Child Relationship Updates
Commander Vehicle and DoorOptionTrigger are now child actors of the BP_TS_Interactive Actor. References are updated on Gamecontroller and WBP_DoorOptions. All Interactive actors in the level have had their box collisions set to original sizes. Interactive UIs are also now uniform and set to 20 on the Z axis. Indicator Components of each Interactive are placed to the closest position to the target.