Smoke, Level Sequences, Dialogue UI, & Spray Trigger Functionality Updated
Smoke added to Sedan that does not go away when the fires are put out (cascade particle). Level Sequences Updated for hoses visibility. Spray Trigger Functionality changed to automatically add the scores as long as the player stays within the trigger space instead of having to go spray back and forth to make it work. Images added to Dialogue UI and Functionality has been updated but not completed.