Morning Updates
Added Backhandline visibility to false on Adjust FrontDoorHose Level Sequence so it will disappear. Exchanged the Escape Button Image for the new one. Edited the functionality so that the Open Nozzle will not appear if you are at the Engine Handline and will only appear at the other handlines. I also added a function to disable the collision and override generate overlaps for the Engine Handline the first time you interact with the Open Nozzle Button so you won't have access to it once you use the handlines. I also added additional functionality for setting the FPSPawn to hidden whenever you interact successfully with an interactive.